"7:20am.
```
normalize: sum
 | strategy normalizingSum |
 strategy := sum collect: [ :each | each max: 0.0 ].
 normalizingSum := strategy sum.
 ^strategy collect: [ :each |
  normalizingSum strictlyPositive
  ifTrue: [ each / normalizingSum ]
  ifFalse: [ 1.0 / Action size ]
  ]
```

Here is `normalize:`. It works. Let me as the question.

https://stackoverflow.com/questions/54966247/how-should-existing-library-classes-be-extended

7:45am. https://stackoverflow.com/questions/4869850/best-way-to-extend-pharo-smalltalk-class-behavior

Ah, yes extension methods.

7:50am. https://tel.archives-ouvertes.fr/PHARO/hal-01483756v1

I remember something about extension methods, but didn't that have something to do with packages rather than classes?

My memory makes no sense here.

8am. https://arxiv.org/pdf/1708.01679.pdf
Scoped Extension Methods in Dynamically-Typed Languages

...Ah, wait, maybe I am thinking about this all wrong?

I am really focused on the `Array`, but what if I made the `normalized` method for it and then moved it out into a different package?

So far I've been thinking along the lines on how to make a method and then move it in. Maybe the opposite is the point of extension methods.

8:20am. Ok, I'll make a `CFR-Extensions` package and extend Array with normalize there. Now let me rest for a bit. Damn, I am tired."