Though I've been keeping it together, just like in early 2017 when it took me a long time to just start work on Spiral, I am truly experiencing a lot of regret over my decision to move to Pharo. I'll do it eventually, but I need to coral my emotions into line first.
These kinds of changes are what separates the truly elite from the pretenders.

There is always the urge to rush to the goal, but the right thing to do is to take the time to do things properly. Right now every little thing bothers me.

I am really throw away a lot of my principles by making the move. That is what makes this so painful.

Going from F# to Spiral, I had to throw away so much of my time and global type inference.

And going from Spiral to Pharo, I am going to have to throw away performance and various conveniences.

The more time goes on, the more I seem to be throwing things away. How suitable. The irony is killing me.

Even simply mastering Pharo as a user won't cut it. Making a ML library turned out to be a difficult language problem in the end. There is just no way that causing the Singularity will be any exception to this.

Lispers think they are hot shit with their macros, but the scope of the task will be enough to crush even them. None of them could come up with Spiral. I had to do it in the end.

Metalearning and metaprogramming will turn out to be one and the same in the end. The closer I bring them together, the closer I will get to actual power in the real world.

It is for that sake that I sacrifice my time and my life in this world. Bargaining with the universe will make results, only harsh training will.

I really do like a lot Pharo quite a bit. The keyword messages are sheer genius - they make the language incredibly readable. And the IDE experience is beyond anything I've ever seen - it is at the level of being a principle. Objects too are a good basis for a language - sending messages gives about 80-90% of the benefits of actual pattern matching.

But the decision to just leave F# again and to leave Spiral is just very harsh. To change programming styles is very harsh.

Pharo does not have much in the way of libraries - pretty much anyone who would pick it up would do it for principled reasons much like myself.

I will dwell on it, agonize over it and keep the decision in my mind until the motivation builds up. I'll do light programming until I muster the drive to do heavier loads.

It is just as well that today there was a power outage for 4h. Let me chill a bit and then I'll try doing the next example in Pharo.

I think I will make this repo my journal after all. Spiral is me, and Pharo isn't just yet. When I truly start breathing and thinking Pharo, I will make the switch.