"8:40am.
```
initalize
 fromIdTemp := Array new: numSoldiers.
 fromId := OrderedCollection new.
 self initFromId: 1 soldiers: numSoldiers.
```

```
initFromId: field soldiers: soldiersLeft
 field <= numFields
  ifTrue: [
   0 to: soldiersLeft do: [ :i |
    fromIdTemp at: field put: i.
    self initFromId: field + 1 soldiers: soldiersLeft - i.
    ]
   ]
  ifFalse: [
   soldiersLeft = 0
    ifTrue: [
     fromId add: fromIdTemp copy.
     ]
    ]
```

Now how do I initalize `numFields` and `numSoldiers`? Lexical provides the optimal solution to all those questions, yet here I do not have the answer. I also need to take care of initializing the players.

At the end of this there will be a lot more code coupling than I bargained for.

8:45am.

```
new: numFields and: numSoldiers
```

I guess I will add this to the class side of `Main`.

But now to set the instance I will have to add accessors.

8:50am. Ok, that did the trick, though the code is full of holes compared to what it would be if I had lexical scoping at my disposal.

Let me do the agent next.

I am not sure if I will even want to test this when I am done. It just such a mess.

Yeah, objects work well for collections whether that be of data or raw methods, but I find that dealing with raw programs leaves much to be desired.

Focus me. Let me do the `Agent`.

8:55am. No enough. I had enough of this. There is no point to writing a program that is not worth writing.

9:05am. I am sure of it now, Smalltalk is just not for me. Even though there might be advantages to live coding, I just can't bear the heavy hits I take everywhere else.

9:15am. Let me chill a little. I need to really figure out how I am going to do things next.

If it is just visualizations rather than GUIs I might be able to deal with it with just F#. Data visualization is a necessary requirement, but GUIs and persistent objects I will just find a way to work around.

Going with F# will make it a lot easier to interface with Spiral and save me a lot of time with rewriting the language yet again.

It is true though that I have a burgeoning interest in dynamic languages and VMs.

I do want to solve the problem of making graphical interfaces for its own sake.

9:20am. http://www.codeconscious.com/rebol/

Rebol is one language that never took off. I am not really interested in looking into it at the moment, but I'll keep it in mind.

9:30am. Well, let me chill for a bit. I do not feel like doing anything right now. I could be doing the CFR examples in F#, but mental energy is too low.

9:35am. I am going to have to modify my plan a little. I'll go with F# and CFR. This will give me something that actually has a chance a working on poker. Then I will build up from there.

https://clementbera.wordpress.com/page/2/

This guy has a lot of stuff on VMs.

https://files.pharo.org/books-pdfs/booklet-CallStack/2018-01-23-CallStack.pdf

I should read this."