I think I will just call it an over for the rest of the day. I still feel aimless, but slowly my mental state is normalizing. I still have the hunger to get better underneath, I just need to find a different path to get to it.
Rather than studying, I feel like programming just a little.

Ok, so I half wasted a week. There is no reason the other 3 for the month need to end up the same way.

There are various ways I can go with regards to interactivity, but all of them are very difficult.

I feel like in that aspect, I'd do well to finally make an in depth study of existing compilers and their runtimes.

It would be great if I knew how ML compilers worked under the hood and that includes F#. Same for Pharo's VM. I am thinking that what determines whether a language is static or not is not the presence or absence of a type system, but the degree of dynamism it has underneath.

Julia is spreading that nonsense that types make things fast, but if a type of a function is `a -> (a -> b) -> b` there is not much info there for the compiler to optimize that. Types only make things faster if they are primitive. I suppose the story might makes sense if you've only programmed in Matlab your whole life and think C is fast because it is statically typed.

If I want to move forward I have to accept the fact that dynamism is a necessity for expressiveness. Dependently typed languages will be pretty dynamic underneath.

I have a resurgence of interest in type systems because in February, F# really proved its worth. After 9 month of programming in Spiral, F# just feels so fresh and easy. It is amazing to now have to think about correctness in terms of both compile-time and run-time aspects. Spiral's features are just a waste of time if I am doing prototyping or programs that do heavy boxing.

I just want to go further.

I want to draw out more of that ease of programming that F# provides. More type system features could be provided without impacting user experience.

Haskell would be a great language if it were not pure or lazy.

I want the next generation ML in my hands today. I want it to provide live coding experience that Pharo does.

I do not think that live coding is incompatible with having powerful type systems. At the very least I can't think of any reason why it would not be.

So I should know what I need to do.

I feel like I've attained a real power in the making of Spiral. I've pushed the low level aspects of programming to its very limit.

Now I need to do the same with dynamism.

I need to master extensional typing and make the next language after Spiral a live one. Not just speed and power like Spiral, but absolute ease of use and persistence as well.

A home for my agents to live in.

I won't run or take shortcuts anymore. I'd hoped that if I completed Spiral and went to deep learning I'd just be able to miraculously make it work using the same methods others have used.

Too naive. I am just as clueless as anyone else with regards what those algorithms are doing.

I need to open the box.

What I will aim for is to be so good at programming that my skill will spill into math on its own.

I will bring the light of reason to computational learning.

Working on data visualization for CFR algorithms while I study how ML compilers are implemented under the hood should be a good first step towards that.

Once more I need to put my pride on the line and become a real power user. I need to get familiar with my favorite tools' internals. It won't just come to me, I need to seek that insight out myself.