"https://fslab.org/report/
5:30pm.

```
(*** hide ***)
#I "../../../bin"
#r "XPlot.Plotly.dll"
#r "XPlot.GoogleCharts.dll"
open XPlot.GoogleCharts
```

Passing references to FSI is still as annoying as I remember it. Shouldn't this have been solved in the last 2.5 years?

Also I really messed up. I should not have put all those FSLab packages into the testing folder of Spiral.

Whops. I am going to have to make a new project in order to fix this.

5:45pm. Oh, lol. I expected a window to open, but instead it sent it to the browser. Maybe I should just go with that.

6:50pm. It seems the FSI issue with sending references had still not been fixed. Well, fine.

Let me stop here for the day. Right now I am just browsing the F# sub. Tomorrow I will play around with data visualization and do the next problem. Either way, I can only count on CFR to bail me out of the pinch I found myself due to deep RL being crap.

I won't let this beat me. I will take the best parts of Pharo and best parts of F# and Spiral and fuse them all into one language. Maybe that is my destiny. Maybe that is why I can into this world.

I really have a huge advantage in that CFR is a tabular kind of method and is proven to work.

With my skills I won't have much trouble scaling it and speeding it up once I bring in Spiral.

But interfacing with those sites...

Yeah, that sort of thing falls into `IDE issues`.

7:20pm. https://www.reddit.com/r/ProgrammingLanguages/comments/ayecs4/spry_redrebol_smalltalk_in_nim/

"> I’d kill for a “modern” Smalltalk that felt as native as any first class IDE/editor experience.

Personally, I'd want a powerful statically typed functional language with the IDE experience of Pharo. I really did my best to get into Pharo in the last week after reading books on it before that, but going from F# and Spiral to Pharo is just too much even for somebody of my own determination. Pharo really looks great on the surface - it is extremely readable, but it is a massive pain to write once one sits down to actually do it.

To move to it I'd need to give up a lot of very dear things to me:

* Let bindings. (Smalltalks declare uninitialized vars and assign to them instead.)

* Built in data structures like tuples and records.

* Pattern matching.

* Lexical scoping.

* Partial application.

* Indentation sensitive syntax.

* Global type inference. (Coming from F#.)

As a functional programmer, the first four hit me really badly. It is just too depressing. It is a massive downgrade in language expressiveness from what I am used to and no, the great IDE does not make up for it.

I do not really see any reason a ML-like language could not have those features, but I guess it would be challenging to retain Smalltalk's open nature. I do not suppose there is a Lisp/Scheme out there with a Smalltalk-like IDE focus?

Or anything whose language design is not permanently stuck in the 80s? Maybe I should give Red a try..."

7:30pm. https://www.reddit.com/r/ProgrammingLanguages/comments/ay8lg1/red_language/

Maybe I should. I saw the language before and read a bit on it, but never tried programming in it.

Well, I'll consider it tomorrow. For today, I am done."