"1:10pm. Let me visualize the flow and then I will do some Red.
I suppose that example counts as doing some Pharo. Damn, I am so screwed.

Let me get familiar with the charts. I really need something real time, like what Pharo offers. Static charts for static data will not cut it.

1:55pm

```
https://developers.google.com/chart/image/docs/gallery/bar_charts
```

Now how do I set the options for this?

1:55pm.

```
let paper =
    [0.2;0.3;0.3;0.2;0.1]
    |> List.mapi (fun i x -> i,x)

let scissors =
    List.map2 (fun (_,a) (i,b) -> i,1.0-a-b) rock paper

let inputs = [rock;paper;scissors]

let chart =
    inputs
    |> Chart.Column
    |> Chart.WithOptions(Options(isStacked=true))
    |> Chart.WithLabels ["Rock"; "Paper"; "Scissors"]
    |> Chart.Show
```

Oh finally figured it out. Ok, great.

Now with this I can visualize the trace of the strategies.

After I am done with this, I am going to have to do this in Pharo as well.

I have to say, after experiencing the open nature of Pharo, the closed form of the .NET libraries that I cannot go to does not sit right with me.

2pm.

```
let print sum =
    {sum with strategy_sum=average_strategy sum.strategy_sum; regret_sum=strategy sum.regret_sum}
    |> printfn "%A"

print (fst players)
print (snd players)
```

Let me back this up here.

2:20pm. The charts for RPS are interesting.

They take forever to display though.

2:40pm.

```
let trace_strategies = Array.init actions.Length (fun _ -> ResizeArray())
let trace_regret = Array.init actions.Length (fun _ -> ResizeArray())
let labels = Array.map (sprintf "%A") actions

for b' = 1 to 300 do
    for a = 1 to 100 do train players

    (fst players).strategy_sum
    |> average_strategy
    |> Array.iter2 (fun (ar: ResizeArray<_>) b -> ar.Add(b',b)) trace_strategies

    (fst players).regret_sum
    |> strategy
    |> Array.iter2 (fun (ar: ResizeArray<_>) b -> ar.Add(b',b)) trace_regret

let show (trace: ResizeArray<_> []) =
    trace
    |> Chart.Column
    |> Chart.WithOptions (Options(isStacked=true))
    |> Chart.WithLabels labels
    |> Chart.Show

show trace_strategies
show trace_regret
```

This way of doing things really sucks. Like really. I am never going to get anywhere like this.

Also the behavior of this code is baffling.

When I just run it, it does the right thing, but if I run the loop a few more times by sending it to the Repl and then show the charts, the thing does not recognize the changes. For some reason no matter what size the ResizeArray is, it the charts always print it in steps of 300. That is completely bizarre. It has to be some kind of optimization gone mad.

2:45pm. I give up. I cannot even imagine why this bug is happening.

But basically, though, I do not need the visualizations.

Supposing I have the CFR agents trained to perfection and parallelized to the extreme on the GPU.

Suppose I somehow succeeded at that. It should be doable for me.

Then when the time comes to let the agents play in the real world, I am going to have huge difficulties.

Interfacing with the Outside is like the reverse interface problem. I essentially have to map the screen to an interface that the agents can interact with.

And here I am having difficulty with some shitty charts.

I've been hung up on this for 30m.

Let me try Pharo again. I want to do the problem in it after all. For nested functions I am simply going to explicitly pass in the arguments where needed. For the rest I am going to set the parameters like in that example.

I am not going to do Spiral in Pharo.

Rather what I will aim for is to pare down the VM to its essentials and then rewrite it into something better. Then I will implement Spiral in terms of that. Though I hate the simplicity of the language, Smalltalk is the pinnacle of IDE design.

Red is interesting to be sure, but I am not interested in just easy GUIs. I am interested in what the language is about.

3:20pm. Let me commit here. I am just staring off into space now.

Let me do some programming.

This is some vicious downswing. But ultimately, I am going to get use to Pharo and put it under my thumb. That will be the day of days. I am not going to be as useless as I am now forever."