Yeah, I am screwed. It does not matter if I throw in the towel. Literally nothing will come close to the Smalltalk IDE experience. And now that I am fully into the 'make interfaces' phase of my plan, even if I give up, my goal is just going to drive me straight back in this direction. It is nice the Red can make GUIs easily, but I require much more than that. I require all that is offered. There is not a Lisp or a Scheme that offers the same IDE experience either.
I did not mean to just fall into this part of the plan, it just happened when I decided to give up deep RL for the time being. There is really a damn lot of work to go here. Even if I do it full time for the next 3 years, I am not sure I am going to make it.

Are you watching this future me? All of this will be for you and nobody else.

It would have been great if Spiral was enough. In the right timeline I would not even have needed to make even it.

There is no helping it. I will do what I can for as long as I can. If I cannot complete the quest, somebody will have to pick up the torch after me. I can't trust the research community to do the right thing. The only ones who deserve to cause the Singularity are those seriously aiming to enter the self improvement loop.

Looking at that path ML is taking, I do not think OpenAI and Deepmind are going to be able to do this.

In a sense, the way I was trying to make ML work in 2018 is just a lesser type of what they are doing. I've been seriously hamstringing myself by doing all my work in Spiral. If it is just research Spiral is a lesser experience than Python. And to make this work I require a lot better experience than even Python.

That will be attained by bending Pharo and Cuis to my will.

To do that, there is no other choice. I simply have to immerse myself into it. Today I will make another attempt at doing what I failed yesterday. I am going to have to rewrite it and impose some sort of convention on how objects are constructed.

During the night I've come to an acceptance that if I want to program in Pharo, I am going to have to leave most of my skill behind.

Well, let me chill a little and then I will start.

Starting from here, I will forget the things that I have done in the last 4 years and pretend I am starting afresh.

For all its flaws, I think Pharo is the purest incarnation of a dynamic language, more so than anything else. Despite all its flaws, if anything is going to save me, then it will be this and only this.