It has been a while since I posted here. Basically, things are not going well. I've put in like 3 weeks of effort into Pharo already, but it is just not to my taste. I can't get used to its pauperous simplicity.
This is the worst, just the wost.

God, I really hate when a language makes me work so much. What a waste of time. Please me give my life back.

Well, I am being overly dramatic. To be sure, I could get around the latest problem with some effort, but I am not sure I want to.

Everything about my experience of the past 3 weeks essentially paints Pharo as a beta-quality language. It is amazing how bad things are considering it has been in development for over a decade now.

The kinds of bugs I've been running into are something that is indicative of the integrity of the language itself.

I sort of threw in the towel a week ago, then picked it up, and now I am considering throwing it in again.

Visualizations or not, this is who I am.

Back in 2015 it would have made more sense for me to just go to Python, but I stuck with F#. I spent 2 years working on Spiral so I could stick with my own style and not have to go to Python.

Maybe other people could throw it away, but I am definitely not the sort that could. If it was Scala or something I could manage, but not this.

So how, just how do I do this?

I sort of understand what I need to do - find ways of visualizing large quantities of dynamic data.

How do I simplify the whole UI creation stick?

I want more than the Elm pattern. I want more than F# offers. I want to rise beyond the level I am currently at. I want to optimize the workflow.

https://youtu.be/Vcoy5gwUvOA
Nomads do not build Cathedrals

The guy giving this talk is exactly right about the importance of the platform.

Just what do I need to do to start operating at that level?

...

Well, before I resume CFR, I want to study the Red language. I want to see what the thing is about. It has been on my backlog for a while.

Right now I am just so completely lost.