I've chilled for an hour already so let me finally start. The challenges ahead of me are so numerous. They weight heavily upon my heart.
https://www.microsoft.com/en-us/research/publication/higher-order-type-level-programming-in-haskell/

Some of the recent developments in Haskell are genuinely interesting to me as an programmer.

Spiral is just one of the manifestations of a language with dependent typing. Eventually, the design of a language that uses types for ease of use will manifest itself. Spiral is brings out the performance aspects of it.

It is really a pity that Idris is a pure language.

Back in 2017 I considered using Haskell and Idris in order to write Spiral, but went with the more familiar F#. In hindsight, that was the correct choice. Functional purity would have made memoization significantly harder to use. With the benefit of experience, I do not see Haskell as that good as a compiler language anymore.

They guys are literally behaving as if they get some sort of prize for avoiding mutation.

Meanwhile, in Pharo there are many places that actually require immutability and the language can't see it.

The ideal style that takes advantage of mutation selectively is something I am using right now. As long as the language encourages the style, the correct actions will come out on their own.

Dependent typing is the future. It will be needed for the kinds of VMs that I desire. It will be needed for the kinds of languages that I want.

Type level programming with constraint solving is just too good of a deal to pass up.

If in the future there is a serious effort to make the next great ML (right now the Successor ML project seems to be going at a snail's pace) I will be there to give a hand.

I'll make up for my failure with Pharo that way. I hate it when I cannot follow through with my intetions, but going to Pharo was just stretching it too much.

What I want to do with interfaces will take a lot more effort than I'd prefer, and it will take innovations in interface technology on my end, but it will fall into place eventually.

I do not how, but I will find a way to interface the agents with the outside world elegantly and safely somehow.

I do not think I will be able to do this task without writing DSLs or relying on Spiral. I need to unearth so much.

But enough ranting.

That is for the future. Though the scale of what I want to do is enough to crush anybody, I will go with the simplest steps possible.

Let me finally do the Kuhn poker example using CFR. It is time to clean up the paper.

There are several points in it that I do not understand and it will take some effort to get through the paper.

It is time to do it.