"4:40pm. 24/38. I know what FSICFR is about without even seeing the code...but the devil is in the details I am going to have to rely on the paper to elaborate.
5:45pm. Though I've went over the section, I can't really say that I've been focusing on the sections itself. I've been getting lost in daydreams.

But at this point in my programming career, simplifying the nasty code like the one from the paper and turning it into a functional work of art is my bread and butter.

Well, at any rate, I will stop for the day here.

Today I did really good work on wrapping up Dudo. Tomorrow I will deal with FSICFR. And after that, I'll be almost done with the paper. I'll clean them up a bit and make a post on the RL sub.

5:50pm. Programming is at its best when it is like this. Every day one learns something, makes a piece and then moves on.

That is the way it should be."