"9:15am. I'd be lying if I said I slept well last night. It was one of those times when obsession grips me.
Tonight I did something I never did before - I actually tried figuring out how type inference should work for a dependently typed language. My recent failures with Pharo and those minimalist ML threads are the spark.

Here is the rub - I understand how to do inference for core ML + typeclasses. I even understand how I could fit such a language on top of Spiral and have it take the burden of doing type inference manually. This is quite a lot even on its own. I could make a ML with a GPU backend at my current level if I wanted to.

But I want my dependent lists so I can implement Spiral's tensors. I want all that an ML would offer, plus extra. If I leave dependent typing on the table because of lack of understanding I would be less as a language creator.

9:20am. Really, nothing went well apart from Spiral in the past 4 years. Spiral has the spark of divinity in it, but I was wrong to try and make it a library language. There is no dynamic language worthy of such honor.

After seeing what state Pharo is in it would be enough to dissuade anybody of dynamic typing. I've run into crazy bugs all over the place even before the recent batch. Remember that cross dragging bug where moving a cross with the keyboard caused it to not get printed correctly? Or that time I managed to corrupt the image and for some reason the letter `h` got printed one to the left - usually over the letter in front of it?

I should not have to explain to the community why they should start looking for ways avoiding mutation related bugs.

9:30am. I've come to a resolution during the night.

I really should learn type theory before the Singularity happens rather than afterwards. Understanding programming and math at a higher level is in fact a prerequisite to making it happen. I have my causality the wrong way around.

The time to learn dependently typed programming is now.

Not just so I can fit a ML with dependent typing on top of Spiral. But also as an exercise in proving the properties of ML algorithms. It is time to bridge together my understanding of programming and math.

Visualization is a failure. I wanted to go down this route not just for its sake, but to also take care of interfaces along the way, but it is not like I have some idea of how I am going to take care of visualizing traces of millions of parameters. Also infinite ranges of parameters that are probability distributions. The only language I could even attempt to do this sort of thing would be Pharo, but it is far below the threshold.

Seeing Pharo as it is today reminds me of Haskell in a way - what made me drop Haskell were the laziness related space leaks. I was trying to debug space leaks in my own programs and the IDE was not working because of space leaks in the compiler. In Pharo, the issues are all mutation related.

As a principle, one should not program in a language where fixing the fundamental issue that it has would require making a whole new language.

9:40am. I will do CFR and master the Elm pattern on the side. I'll go with the poker plan as a concrete motivator.

But even though it might not have been my intention, mastering intensional typing opened a lot of doors for me. If I can level myself up here, not only will I crack the credit assignment problem in ways that will be irreproachable, but I will also make the next great ML. Because of Spiral, I have everything I need to make a fast dependently typed language.

This is something fundamentally worth aiming for.

9:45am. Trying to implement the work of other people really has not been working out for me as a strategy.

If I can overcome the whole 'not good at math' thing I could really go places.

I got an enormous amount of practice since my last (failed) attempt at mastering dependently typed languages back in 2016. This time I really could do it. If I want to try Coq again, this time I will play it smarter and figure out what it does under the hood first. It should reduce to a small amount of primitives.

Principled exploration will always beat doing random things and hoping they stick.

The trouble with doing random things is that can fail severely and anything at all could fail to stick. This is killing me in ML. This killed my last attempt at going through the Software Foundations book.

10:10am. Ultimately, I just cannot make the move to Pharo. F# is hardly the pinnacle of programming, but compared to Pharo it is on an entirely different level in how easy it makes reasoning about programs. I can throw away the work that I did, but I just cannot make a move that would weaken my ability the degree that Pharo requires."