"11:55am. Let me start this thing. Let me see if I can get some focus going. My sleep was bad tonight so I won't be at full power today. My eyes have been feeling tired ever since I woke up, but regardless - I must go on.
12pm.

```
Note that the only relevant information to the doubt/accept decision is the past two claims. No claim prior to those have any bearing on the current decision, as the die has been rerolled and is without memory.
```

Why the past two claims? Should it not be just one?

```
All a player needs to know is the current claim and the constraints on that claim. Note also that the only relevant information to a new claim decision is the roll the claim concerns and the constraints on the claim, i.e. the previous claim if existent.
```

I think this is just the case of the authors using a separate bool to state whether the claim exists or not.

I'll assume that this is the case.

12:25pm. I getting distracted by the FF remix.

Focus me.

1:35pm. I am focusing.

I think I sort of get what the algorithm is about. The second half of FSICFR is essentially backpropagation.

The first half is a straightforward graph problem which I am not sure how to tackle.

Here is the issue I am having.

```
let nodes = [1;2;3;4]
let nodes_outgoing, _ = List.mapFoldBack (fun a s -> (a,s), a :: s) nodes []
```

```
val nodes : int list = [1; 2; 3; 4]
val nodes_outgoing : (int * int list) list =
  [(1, [2; 3; 4]); (2, [3; 4]); (3, [4]); (4, [])]
```

This is the standard CFR form that does an exponential number of visits. This is what I've been using in the other examples. To turn this into backprop, what I need to do is reverse the order of edges from `nodes_outgoing` to `ingoing_nodes` and then iterate them in order.

I am actually not sure how to do that actually. The issue is that I do not want to assume the ordering - though for Liar Die it is given. I want the graph traversal to be done in a generic fashion.

The only assumption I want is that the initial node is known. The rest can be in any order...

```
[(1, [3; 4; 2]); (3, [4]); (4, []); (2, [4; 3])]
```

How would I transform this?

Well, I think I could manage the transform to the incoming edges. All it would take is some simple memoization.

2pm.

```
open System.Collections.Generic

let inline memoize (memo_dict: Dictionary<_,_>) f k =
    match memo_dict.TryGetValue k with
    | true, v -> v
    | false, _ -> let v = f k in memo_dict.Add(k,v); v

let nodes = [1;2;3;4]
let nodes_outgoing, _ = List.mapFoldBack (fun a s -> (a,s), a :: s) nodes []

/// Shuffled
let nodes_outgoing' = [(1, [3; 4; 2]); (3, [4]); (4, []); (2, [4; 3])]

let outgoing_to_ingoing initial (graph: Dictionary<_,_>) =
    let nodes = Dictionary()

    let rec loop node: HashSet<_> =
        memoize nodes (fun node ->
            List.iter (loop >> fun x -> x.Add(node) |> ignore) graph.[node]
            HashSet()
            ) node

    loop initial |> ignore
    nodes
    |> Seq.map (fun kv -> Seq.toList kv.Value, kv.Key)
    |> Seq.toList
```

As usual, recursion comes naturally to me. If there is something I've gotten good at, it would be recursion.

Let me test this. First let me get the conversion function and then I will figure out how to do a sort on top of that.

2:05pm.

```
val it : (int list * int) list =
  [([3; 1; 2], 4); ([1; 2], 3); ([1], 2); ([], 1)]
```

Here is what I get now. Now how do I sort this given that labels might be permuted?

NNs comes with their ordering attached, but here I am not sure.

Let me take a look at resources on topological sorting. The constraint that I have to satisfy is that the nodes in the list are greater than the one on the right."