"3:50pm. https://arxiv.org/abs/1903.08254 Efficient Off-Policy Meta-Reinforcement Learning via Probabilistic Context Variables
This is pretty interesting.

3:55pm. I'll read it later. Let me go back to studying the CFR paper. I'll take a look at this once I am done with CFR. And once I am done with CFR, I'll move on to the Little Typer.

25/38. There is a lot of stuff here I do not understand actually. The algorithm is substantially different than the standard CFR.

4pm. Actually, the first 25 lines of this are rather clear. I have no idea why I was afraid of it.

4:15pm. https://www.reddit.com/r/MachineLearning/comments/b3gdyi/d_what_are_some_good_life_lessons_from_ml/
"its really difficult to reason outside your experiences"
"This. I was explaining ML to a family member a while back, and interestingly this was their main takeaway/observation. It hit me like a brick then, and it hits just as hard now that you've said it."

Honestly, the pseudocode now that I stared at it for 20m is so readable. It is amazing. I feel that my odds of implementing it went to 100% now.

Now come the code. I think I will leave the actual implementation for tomorrow.

https://mangadex.org/title/31477/solo-leveling
https://mangadex.org/title/32020/iron-ladies
https://mangadex.org/title/33538/kengan-omega

I'll save these for later. For now, let me focus for a bit longer.

4:45pm. I've planned out most of the stuff. I understand the pseudocode and have gone over the code. I won't focus too much on the code, as it is highly specialized to the example in the paper. It is really low level.

4:50pm. I think I will call it a day here.

4:50pm is not the time to enter a deep a flow state that will last hours. Trying to do it here would only endanger my health.

All in all, I haven't done too badly. There is plenty of time to do this properly and then move on to the next thing.

Maybe I will check out that paper when I get through CFR. I am done with policy gradients, but I might end up using NNs in a unsupervised manner for the sake of performing dimensionality reductions and such. Or maybe even the stuff in that paper. Whatever I do, my next attempt is going to be more Bayesian in nature.

4:55pm. But forget about that. Mastering dependently typed programming and along the way learning how to analyze these algorithms is what really matters.

That PCV paper might be hot now, but tomorrow another thing might be good. I can't just bounce around like a pinball from algorithm to algorithm. I have to be able to grasp the insight on my own.

In the next 3 years, I will wrap things up.

Things are not going as planned, but in terms of life expectancy, I have the time to see things through. The long term is scary and unexpected tragedies might hit and make me stumble, but there is no immediate danger and it is not likely either.

Realistically I could spend half a decade or more of just doing it like this, just learning.

So I am going to do it."