"10:15am. Ah, I really was foolish.
Looking at DIPPL once again, I see that chance and response nodes in CFR are similar to sample and factor nodes in a probabilitic language.

11:30am. I am still thinking about it.

This just goes to show how hard this really is.

11:55am. I am still thinking about it.

Basically, I am going to do it, and it will be in the simplest way possible. I am going to memoize the game along with them the nodes. The players will be part of the key.

The game nodes will be a lot larger than the action nodes, but there is just no way around this. Not for this kind of algorithm. FSI trades exponential time for exponential memory.

12pm. Let me just push the off switch on my brain for a bit if possible.

12:05pm. FSI is way harder than I thought it would be. There are just so many design constraints to work through that it boggles my mind. But the simplest implementation should not take more than a few hours.

Hopefully, Libratus and the other CFR schemes won't turn out to be as difficult as this."